Remove eth_retry dependency
Eth_retry introduces as dependency Brownie. This is a bug, since Brownie
is only used in eth_retry tests. Due to this bug, a tests/conftest.py is
installed in your virtual environment when you install eth_retry, and
this conftest.py tries to import Brownie, so this results in error if
the project where you are importing eth_retry (and therefore,
multicall.py) has no Brownie installed.

This behaviour has been observed in eth_retry versions 0.1.17 and
0.1.18.